yin parser BUGFIX duplicated parsing of augment's common substatements
Augment's substatements description, reference and status were parsed
twice (or tried to be parsed, but the nodes were not there already) -
first on parsing augment and secondly when resolving augment. Instead,
when resolving augment, we only need to inherit config property from
the target node, this cannot be (and is not) done when parsing augment.