json printer BUGFIX improve check of last node
JSON puts comma between nodes, but it cannot be after the last node.
So far, the checking of the last node were failing when a list's or
leaf-list's instances were mixed with leafs (or containers, etc.).

The commit adds a standalone function to correctly detect the last
node in case the list's and leaf-list's instances are processed at
once.