xpath NEW keeping the document order
Document order is modelled based on the
data document, not possible with schema
(lists and their children could not be
distinguished). For this purpose
a general sorting function is used,
will change from select sort to bubble
sort to handle sorted or almost sorted
sets effectively. Also, namespaces are
unsupported now.