xpath NEW evaluating all operators
Execution of comparison and logical
operators in a function, not implemented
yet. Set and node set unified, memory waste
negligible and not worth all the casting.
All the expressions separated to a function
for each, there were 2 expressions evaluated
in one function before. The recursive
function calling overhead should not be
significant and it should make those
functions incomparably more comprehensible.