parsers FEATURE support get-filter-element-attributes extension
Kind of fake support - nothing special is needed for this extension,
it just allows some XML attributes to appear and libyang already
keeps all attributes. libyang does not handle the defined attributes,
it is up to the caller. This change actually make silent warning
about unknown extension in the case of ietf-netconf schema.