xml FEATURE provide public API to a simple XML parser
Allow caller to build XML trees on its own and traverse the XML tree.
Function to manipulate with the tree are not provided.

Data tree parser was extended to accept XML tree besides the XML as
string. Also XML dumper was extended to accept option to print only
a specific part of the element (opening/closing tag, attributes).

This way a NETCONF app will be able to handle RPCs, Notifications,
hello messages which are not described in a YANG schema.