XML printer CHANGE remove unused namespace definitions
When parsing XML, do not keep namespace definition that are not used
by attributes. When printing XML format, namespace is created (according
to the same rules as in the case of JSON, but in XML format with using
default namespace). This unifies the approach in case of input data
in XML and JSON format.