JSON parser/printer CHANGE support empty elements for filters
To allow expressing subtree filters as JSON data, enable empty
items with 'null' value. draft-ietf-netmod-yang-metadata-02 does
not specify this - it is actually not needed for the way JSON is
currently used in YANG community. However, we want to have full
compatibility between XML and JSON for future possibility of
having even NETCONF operations in JSON.