data tree BUGFIX accept LYD_OPT_NOSIBLINGS option in lyd_validate()
lyd_validate() was running always on the subtree from the provided
node. But since it accepts parser options, it must follow the behavior
of the parser, so by default it now validates all sibling nodes of the
provided node and only if the LYD_OPT_NOSIBLINGS is specified, the
single subtree of the given node is validated.