schema tree CHANGE processing submodules
Data (nodes, groupings, rpcs, notifications) from submodules are now
stored in the main module's data tree to simplify searching for the
data. For the same reason, the import and include array in the main
module are extended by imports and includes from submodules (each
item is added only once to the main module).

This is a huge change and includes also some other optimizations and
bugfixes that appeared with submodules.