context CHANGE use implicit yang module to represent YANG attributes
RFC 6020 defines several attributes in the YANG namespace. libyang
is able to represent (store internally) only data that are somehow
connected with a module in the context, but this was not possible
in the case of the mentioned attributes since there is no such
data model. So libyang has its own (empty) YANg schema to be connected
with these attributes.