CHANGE allow hiding validation issues
In case of forward references, a validation error is actually not an
error and it will be solved later. Therefore, allow avoiding printing
error messages in such a case (message is still created and stored
so it can be accessed via ly_errmsg() and ly_errpath()). Then, a
parameter 'first' in many functions is no more needed and it was
removed. For certain cases (when, uses) the resolving is done in loop
because sometimes more than 2 tries are needed to finally resolve them.