xml data parser CHANGE apply strict option to attributes
In XML, attributes are, by default, without namespace. libyang requires
everything to be in a namespaces of the known schemas. So far, XML attributes
with no or unknown namespace were ignored and warning was printed. With
LYD_OPT_STRICT option, it is natural that instead of warning, error is
supposed to be generated. Without the strict option, warning is still
printed.