xpath OPTIMIZE xpath sorting optimalizations
In release build, all sorting is skipped and
it is assumed that the only XPath construct
that can break the ordering is union, in that
case a simple merge of sets is performed (in O(n)).
To increase chances of detecting cases when
this assumption does not hold, in debug build
sorting is done where the ordering is required,
and if the original set is not ordered
an error is printed.