node set CHANGE return information about found duplicity in ly_set_add()
Changes API.

ly_set_add() does not (and never did) add an item already present in
the set. But so far, such an attempt was just silently ignored and
success (0) was returned. Now, instead of 0 for success, the function
always returns index of the added item in the set. If the item is a
duplicity of an item in the set, the index of the present item is
returned and the item being added is still not added.