context CHANGE special (more generic) handling of internal schemas
keep internal schemas information in a special structure to make it
available on multiple places without need of changing the code if
some other schema will be added in a future.

It also handles internal schemas in a special way in ly_ctx_load_module()
to avoid issues with importing internal schema without revision
specified. In that case the function tries to load the latest schema revision
which make no sense in case of internal schemas.