schema parsers BUGFIX resolving leafref's path
We need to resolve leafref's path in schema, but that is not possible
in some cases, so we have to avoid the resolving in case of:
- typedef
- any type inside a grouping definition

because in these cases, the relative path can point to missing nodes.
This is not a bug since path is supposed to be resolved in data tree,
so in leaf and leaflist instances. We also need the resolved path
only for the schema nodes that can be instantiated in a data tree,
so skipping the path resolving in other cases is ok.