data validation BUGFIX validating inserted list/leaflist instances
When a list/leaflist was constrained by max-elements statement and a
new instance of the list/leaflist was inserted (breaking the max
constraint), the contraint was not checked anymore because the parent
node was still marked as valid. Unlinking as well as inserting a
list/leaflist instance constrained by min/max-elements statement now
adds a validity flag to the parent to force checking for mandatory
nodes (children) during the validation.

Fixes #89