defaults BUGFIX autodelete of empty containers when removing default values
When a presence container was detected, the algorithm was continuing to the
next parent and checking for non-presence container repeated. But this time
the to_free element was still the node under the presence container so in
case of non-presence container as parent of a present container, it wrongly
deleted even the present container.

Fixes #96