schema parsers BUGFIX checking default values for enumerations/bits
YANG 1.1

Since enums and bits can be restricted and the default value inherits
from the base types, we have to check default value in the derived
enum/bits types to make sure it is still valid. In YANG 1.0 the enums/bits
type (list of values) cannot be changed so it wasn;t necessary.