data parsers CHANGE filling data information in case of unresolvable leafrefs
In case of parsing get/get-config/edit-config data, leafrefs are not
resolved, since the target can be omitted in the data tree. So far, the
nodes were marked with LY_TYPE_LEAFREF_UNRES flag. Now, besides this flag
we also resolve the value string to get the value in the proper representation.