defaults CHANGE fix dflt flag in non-terminating elements
When a data subtree is unlinked or created, it can affect the dflt
flag in its parents, so we have to check and fix these flags. Possibly,
this can be done on tree validation, but it should be more effective
to do it immediately (in validation, we would have to through the tree
from leafs to root, which is opposite direction than now).