data tree CHANGE improve lyd_insert(), lyd_insert_after() and lyd_insert_before()
Make the functions more smarter to be able to replace the implicit default
nodes instead of just adding the new nodes next to the previously automaticaly
created implicit default nodes.

lyd_insert() also supports smart inserting of a key into a list - it is able to
put the key into a correct position.