all CHANGE follow conformance mechanism described in YANg 1.1
better detection if the model is implemented based on how it is used
by other implemented modules:

- If the module is not implemented, augments and leafrefs are not connected
with their targets.
- All the modules augmented or referenced by an implemented module are
also implemented
- module can be set as implemented via a standalone API function
lys_set_implemented()

TODO when searching for the augment/leafref target, the implemented revision
of the module must be used despite the revision used (or not) in import.