data parsers CHANGE support out of order list keys
JSON format does not require list's keys ordered according to schema
at all. In XML, the correct ordering is now required only in case of
LYD_OPT_STRICT parsing, otherwise it is able to get the key whereever
and put them into the correct position - keys are still always (in XML
as well as in JSON) printed in the order as defined in schema.