XML parser CHANGE allow empty leaf when it is deleted in edit-config
In edit-config, when a leaf is being deleted (it has attribute operation
in NETCONF base namespace and its value is 'delete' or 'remove'), it
is allowed to have no value even in case the data type does not allow
empty value. In such a case, the leaf is going to be deleted despite
the current value of the leaf in the target data tree.

Note, that this behaviour is possible only for leafs.

Fixes #130