log BUGFIX use schema instead of data node for path reference object
Since data tree can change quite fast and system often reuse previously
used memory, use schema nodes as a reference pointer when checking if the
current error path refers to the same node which is currently used in error
message. It also optimize error path generation when different instances
of the same node (so the same error path is being generated) are used.