log BUGFIX avoid collision of generated paths for log messages
Since we use schema node for identification of currently present path
in an internal buffer for schema as well as data trees, we can have a
collision between the schema and data paths (that usually differ for
the same node). We have to remember the type of the element to avoid
such a collision.