schema tree BUGFIX change 16b bitfields usage according to endianity
in schema nodes we use bitfields in uint16_t for storing flags and
nacm info. But because of magic in grouping node where is changed
number of bits used for each bit field, we have to distinguish between
little and big endian environment to keep the bits on appropriate place.

Fixes #148