logger CHANGE avoid nonsense/duplicated messages when checking type
part of the type checking is (in some case) checking for compatibility
of type changes (of the derived types) with a possible default value
specified in some of the base types. When the base type is not resolved,
it was also reported that the default value could not be checked yat and
it will be attempted later. Since it is actually just a consequence
of unresolved base time, it does not make sense to report it (with
null default value) in this case.