default values BUGFIX LYD_OPT_NOAUTODEL on default nodes
The option is not supposed to be applied on default nodes, because
in the first step we create all default nodes and then they are
supposed to be removed according to the result of the when condition.
There must not be an option to disable this. LYD_OPT_NOAUTODEL is
applicable only to the nodes provided by user.