schemas BUGFIX maintain backlinks in case of removing module(s) from context
In 3 cases, the target module keeps backlinks to the referencing module to
optimize further operations: identities, features and leafrefs. When a referencing
module is being removed (ly_ctx_clean() or ly_ctx_remove_module()), the backlings
in the remaining modules must be checked and updated to avoid pointing to the
freed memory.