schema parser BUGFIX inheriting config-stmt under uses
when instantiating data nodes into final uses (not in another grouping),
the config flag must be inherited bofore resolving the type of the node.
If the type is leafref, we are testing if it points to the node with the
same config status, so the config flag in leafref is supposed to be already
corrected.