data parser BUGFIX getting base type of leafref when it is not being resolved
Even in case the leafref is not supposed to be resolved in data tree
(e.g. in case of NETCONF get/get-config data), we still need to get
base type, so the leafref must be resolved according to schema tree.
In case the leafref was pointing to an union type, this wasn't done
and while no error was printed, the data parsing failed.

Fixes #171