data parser CHANGE keep all empty containers being parsed despite the data tree type
In LYD_OPT_DATA and LYD_OPT_CONFIG trees we add empty containers as default
nodes to be part of accessible tree. In other data types we do not do that,
but we also removes empty non-presence containers when such a data tree is
parsed. A more proper and clean way is to keep such elements, but mark them
also as a default node (so they are not printed by default). It also save some
resources avoiding removing empty container and them recreating it as a
default node.

Fixes #174