yanglint CHANGE include ietf-yang-library data when parsing complete datastore
parsing complete datastore expects to have all the data in input
including the status data. However, The mandatory status data are also
defined in internal libyang model ietf-yang-library, so yanglint is
expected to add these data for successfull validation of the datastore.

This behaviour was already implemented in the non-interactive mode, so
this commit affects only the interactive mode.

Fixes #177