parsers CHANGE make difference between canonical and lexical forms of values
Accept all lexical forms of the input data, but keep them and print them
(i.e. transform them on parsing) in canonical form.

The implementation is not perfect in case of leafref and unions - there
can be a situation when to get the right data type, we need the comparable
(canonical) value, but to get the canonical value we need the correct data
type so it is a cyclic issue. In such a case we do not transform the given
value into a canonical form so it can lead to not resolved leafref.
This issue does not appear if the leafref values are provided always in
canonical form (simply with the same value as its target value printed
by libyang).

Fixes #175