schema parsers BUGFIX instantiating lists from groupings
Do not set config flags for the grouping and inherit it inside it,
the config flag always depends on where the grouping is instantiated.
Then, when no config flag is set, skip some checkings for config
compatibility between parent and its children. The checks are done
later when the grouping is instantiated and config flags are set
correctly.

Similarly, do not check for list's keys in grouping, since it can be
added even by augment only to the finally instantiated grouping and
the list definition inside the grouping will be always without the key
(that's a terribly designed schema, but it can be defined and avoiding
these tests in grouping can save some time)

Fixes #132
Fixes sysrepo/sysrepo#487