Use GNUInstallDirs for destination guessing
Warning: on non-Debian amd64 systems, this will change the installation
directory of libraries, they will be put into
${CMAKE_INSTALL_PREFIX}/lib64/ now. The old versions of libraries which
might still be in lib/ will be left in place, which means that there's
some potential for breakage on developers' systems.

The same applies to PKG_CONFIG_PATH which is changed from lib/ to lib64/
on these systems.

Still, I think that following cmake conventions is a good reason to
follow suit here.