schema trees BUGFIX resolving augments to non-implemented modules
if the augmenting module is implementing, the augment target modules
must be set implemented before it is started to resolve. The problem
is that if the target module is not implemented and the augment target
is from another target, the target does not exists until the mentioned
module is set implemented.