schema parsing CHANGE try to differentiate XPath errors
We are checking XPath expressions (of must and when) already when parsing
the schema. If the mentioned nodes are not found in the schemas, the
expression will always fail. Now we try to differentiate situation, when
it is certain error (the expression refers the node in the same schema
which is missing) and when it is possibly caused by not implementing another
module (the expresison refers the node from another module and this node is
not present). In the second case, we not report the issue as error, but only
as a warning. The schema is then correctly parsed, but the data cannot
appear anyway, since the must/when expression will always fail.