JSON parser/printer BUGFIX anyxml's JSON representation differs from anydata
distinguish between anyxml and anydata in JSON - anyxml allows to have
any XML represented and any JSON node/value pair, anydata allows only the
data that are modeled by a YANG modules despite they are unknown.

anyxml JSON representation is too generic, so we support only the string
value of the anyxml node. Otherwise we are not able to parse it into the
internal libyang structures.

Fixes #200