data tree BUGFIX inserting list/leaf-list instance in rpc/action
Since the order of the data nodes in rpc/action is strict, we are
searching for the correct place where to insert the node. However,
in case of list and leaf-list, it must be taken into account that
there can be already some instances of the same list/leaf-list and
we are actually supposed to put the new node as the last instance.

Fixes #202