data printers BUGFIX printing default data in notification/rpc in explicit mode
nodes in notification/rpc are not config or status, they do not have
the config flag set internally, but when printing default nodes we are
checking for the config status of the node in case of printing in explicit
mode, in which the default config false nodes are supposed to be printed.
The test had to be changed to detect not presence of config false flag instead
of checking for presence of the config true flag (which is not the same since
the nodes in notification/rpc have no config flag set).