parsers CHANGE simplify/clarify lyp_parse_value()
separate 2 different features of the function:
1) leaf parameter is necessary for resolving value and its type (especially
in case of identityref), so we cannot depend on it to get know if 2) the result
is supposed to be stored in the leaf. Therefore, a new parameter of the function
was added to get know if the result is supposed to be stored and the leaf
parameter is now mandatory.