data parsers and printers CHANGE modifying identityref without prefix
in case of the default prefix of the identityref, we add it explicitely
into the string value (kind of canonical transformation). It simplifies
resolving leafref since it expects that the strings are canonical and always
matches. This is not true in case of identityref because the default prefix
depends on the node where the value is placed.