schema CHANGE YANG extensions support
- redesign the support for YANG extensions in schema tree structures.
The extension instances can be stored just in part of the YANG statements
represented already as a structures. To avoid complex change of the currently
used structures, the extensions instantiated in subset of YANG statements
like 'description' or 'default' are stored in the extensions list of their
parents and marked with the flags specifying the correct substatement where
they are actually instantiated. This approach also saves some memory since
the extensions are not very common.