extensions FEATURE allow inheritance of the extension instances
sometimes (mainly in case of flag extensions) it may be useful to
inherit the extension instance into its children. Such an inherited
copy of the extension is not printed as part of the module, but it
is available when working with the schema tree to get know that
the (extension) flag is connected (since inherited from a parent)
with a schema node.

The feature is used in the NACM extensions implementation since it
simplifies checking for NACM extension presence.