schema parsers BUGFIX resolving posponed extension instances
due to a possible reallocation of the extension instances list, the
resolver must be able to get the proper location to store the resolved
extension instance into a reallocated array. Therefore we now pass
the address of the pointer to the list and the index for the extension
instance beign resolved in the list.