schema parsers CHANGE including submodules
Better follow RFCs (mainly 7950) when including submodules
and actually simplify the functionality. Since 6020 was not strictly clear
about submodules, we were allowing to include submodule in another submodule
while it wasn't included in the main module. pyang and 7950 is strict about it
and it is not allowed.

Due to a future changes connected with extensions support and also to simplify
the code, in 1.1 modules all the includes are required in main module and in
case of 1.0, warning is printed, the missing include is added into the main
module and parsing succeeds.