extensions BUGFIX keep the correct module for the postponed extensions resolving
In case of instantiating extension in submodule, we need to keep the
link to the submodule for the postponed resolving of the extension.
The final resolving is done at the end of parsing the main module and
generally used module is in that time the main module. Because it can have
the different set of imports (including the case there is a different
module imported with the same/colliding prefix), we still have to use
the correct submodule where the extension is actually instantiated.