diff BUGFIX comparing ordered lists inside another list with multiple instances
When libyang was checking for changes of an ordered list/leaf-list which is
placed inside another (not necessarily ordered) list with multiple instances,
it mixed instances of the ordered list/leaf-list from all the instances of the
parent list. Furthermore, the internal structure for information about the
nodes positions was allocated for the instances in the first instance of the
parent list, so it could cause segfault.

The patch adds new internal function to check that internal list of positions
information is always found for the correct instance by checking all the nodes
in the parents chain (including the keys of the possible lists).

Fixes #250