context CHANGE order of removing modules from the context
Adding modules into the context guarantee that imported modules are
added before the importing modules. When removing modules, so far we
were going from the imported modules to the importing due to a safe
removal of the augments and deviations. However, due to freeing
extension instances, we need to keep the extension definition in the
imported modules until all the extension instances are freed (so we
need the opposite direction of removing modules from the context).
To allow this, before the module is removed from the context, all the
augments and deviations it defines are now removed from their targets first.