metadata CHANGE support for NETCONF's edit-config's and filter's attributes
The NETCONF's attributes operation (for edit-config), type and select (for
filter) are not defined as standard annotations. To provide the attributes
in data trees in the same way as other standard annotations, we add internally
visible annotations into the ietf-netconf schema in context. This way, libyang
parse the attributes as any other data attributes defined via annotations.