XML parser BUGFIX handling XML elements with mixed content
In case of parsing XML for YIN, we can disable support for mixed content,
since it is not allowed in YIN at all. In case of XML data, it must be
allowed for the case of the content of anyxml nodes, but in other cases,
the mixed content must be detected and reported as error.

Fixes #270