xpath BUGFIX resolving module names in augment's when/must xpath expressions
by default, the module names used as prefixes in XPath expressions are resolved
from the imports of the module of the context node. However, in case of XPath
condition placed in an augment, the context node is the augment's target, but
the prefixes must be resolved in the module of the augment.