schema parsers CHANGE remove unnecessary error message
If the uses cannot be resolved because the grouping is present, but
it is not completely resolved (includes unresolved uses), the bug
is the uses/grouping inside the grouping, and such error is printed.
It is not necessary to print another error about the grouping refering
an unknown grouping, it is kind of false positive error message.