schema parsers BUGFIX use the original module when resolving unresolved items
instead of the origin module of the unresolved item (which could be a submodule),
the main module was used when resolving the items at the end of parsing. This
can cause issues with resolving the imports - instead of using imports in the
origin module, actually the main module is used to search import with the prefix
from the submodule.