parsers CHANGE handling idenittyref's values defined in non-implemented schemas
According to RFC 7950, sec 9.10.2, valid identityref's values must be identities
from the implemented modules. This restriction was not applied in libyang so far.
As a consequence, it changes also handling of the identityref's default values
defined in the schema. We understand it as a direct intention to use the imported
module as implemented. Therefore, if an implemented module refers as its identityref's
default value an identity from a non-implemented module, this module becomes
implemented. This applies also in case of instantiating typedef with the default
value in an implemented module (note that the module where the default identity
is defined is change to implemented, not the module where the typedef is defined).
The rule does not apply:
- when the default value is specified in typedef, but it applies as soon as the
  typedef is used in a data node
- when the default identity is defined in the same module (module cannot be changed
  to implemented by self-reference)