schema parsers BUGFIX remove invalid assert
assert for having main module as a stop in the list of submodules being
parsed for circular check is invalid because it works only for the case
there is one stop in the list (actually in that case there is no item
in the list which is the reason the chack passes - the assert is actually
not applied). In case of some submodule's import represented by multiple
submodules, the import is not yet inserted in the list (but in this case
the assert was applied) and the check fails.