libyang CHANGE allow using different versions of ietf-yang-library
By default, use the new ietf-yang-library revision (2017-08-17) as an
internal module, but allow caller to avoid its loading and instead the
caller can load its own version. In such a case, libyang will provide
context info (ly_ctx_info()) only in case the ietf-yang-library is
loaded in revision 2016-06-21 or 2017-08-17. Other revision can be loaded,
but libyang will not provide the context info in this form.