schema tree CHANGE check for arrays size limitations
some of the statements are stored in arrays with stored size, but
the size is limited (limit is various - 256, 65535,...). In case the
schema contains more than this number of the specific statements, it
overflows and the previous statements information is rewritten. Now,
the overflow is being checked and parsing fails with internal error
and the message announcing which limit was reached.