schema parsers CHANGE resolving uses's augments
augments in uses are  supposed to be resolved all the time, despite
the module is implemented or not. These augments affects only the
uses's own subtree, not other modules. Furthermore, the uses with
augment can appear in grouping which is supposed to be used even
outside (the impoerted, not implemented) module with the applied
augment.