cmake: Do not write to the source directory
We're including libyang as a git submodule in our downstream project.
Once we perform the build for the first time, `git status` in our
internal integration repository starts showing the libyang submodule as
"dirty" because libyang's build places files into the source directory.

This commit ensures that the original, source directory remains
unmodified.