xpath BUGFIX do not remove prefix from identityref values
Identityref does not have any canonical form, internally it is always
stored with its prefix so the string comparison in XPath expressions
should always include the prefix. As a consequence, identity values in
XPath expressions are now limited to be always prefixed, the default
prefix does not apply in this case.

Despite this endeavour to correctly apply XPath string comparison on
identityrefs, the correct way should be to use derived-from() or
derived-from-or-self() functions in XPath expressions.