plugins BUGFIX repeted loading of plugins
There are 2 error cases.

1) The plugin is already loaded by e.g. main application for some reason,
then it was detected by the plugin subsystem as already loaded and therefore
the plugin was not preprocessed and not available for use.

2) Some libc implementation (musl) does not implement reference counting
of the loaded plugins and their dlclose() does actually nothing. In such
a case when the only context was destroyed, all the plugins were supposed
to be closed, but they were not. When another context was created, the plugin
itself was detected by dlopen() as already loaded, but we did not have the
needed information about the plugin functionality so it was not available
for use.

It is solved by detecting if we already have the plugin also in an internal list
of handlers instead of using RTLD_NOLOAD and testing the result. As a side effect,
it extends compatibility, because RTLD_NOLOAD is not available in some older
libc implementations.

Fixes CESNET/Netopeer2#138
Fixes CESNET/libyang#483