Fix issue when defaults aren't stored in canonical form
Sometimes the `dflt` value on a type isn't stored in canonical form (for
example, storing '0xa' instead of '10'). This isn't a problem by itself,
but it makes it important that functions that create default values
properly pass-on the `dflt` flag.

This commit also addresses a secondary issue that stems from the
primary fix. Because the `dflt` flag is now being propogated properly in
more circumstances, an issue crops up when a default value for an
identityref is given in a typedef from a separate module. The fix is to
also search the typedef-defining module to find the module that defines
the specific identity instance.