C++: Do not use CFLAGS as flags for the C++ compiler
The CXXFLAGS and CFLAGS are different variables even though they happen
to accept a common subset of options. It is always wrong to mix them
like this.

This particular usage breaks builds with clang's libc++:

- the build environment specifies CXXFLAGS=-stdlib=libc++
- libyang silently ignores that variable and produces a C++ library
which uses symbol names from GCC's STL
- a user application which links with libyang's C++ binginds fails to
build because it epects, e.g., a std::__1::shared_ptr<Deleter>.

This patch ensures that CXXFLAGS are not ignored, and that the CFLAGS
are not accidentaly passed to the C++ compiler. On exotic setups, this
might require setting CXXFLAGS in addition to CFLAGS if CFLAGS contains
hand-crafted options such as include paths which bypass the regular
cmake package discovery, etc. If the developer wants to pass some custom
optimization options, the correct approach is to set them in both CFLAGS
and CXXFLAGS.