JSON: fix bug on platforms where char is unsigned
ARM defaults to `char` having the same range as a `unsigned char`, and
therefore comparisons for `>= 0` are always true, and the compiler
correctly warns against them.

By first converting the data type to a signed char, we ensure that the
code behaves in the same way as if the `char` was already a signed type.
That's apparently the developers' intention here, and that's how x86
operates.

fixes #580