Fix lyb bit-handling bugs
This commit fixes a couple of problems with the handling of the `bit` type
for the `lyb` file format. Both bugs were only exposed when there were more
than 8 values defined for the bit type.

The first bug was an off-by-one error. If the user attempted to set the value
at position 0, no values would be set. If the user attempted to set
the value at position 1, position 0 would be erroneously set instead, and so
forth. I fixed this by correcting the bit shifts in `lyb_print_value()`.

The second bug was due to the fact that `lyb_parse_val_1()` was not advancing
the data pointer to read subsequent bytes for subsequent bit values. Thus if
the user had set the value at position 0, it would appear that the values at
position 8, position 16, and so forth were also set.

This commit also contains a test modification to verify the fixes. Without the
source code patches, a set of `position-0 position-5` will result in a data
tree with `position-0` not set at all, with `position-4` set instead of
`position-5`, and `position-4` erroneously mirrored into `position-12`.